First try for thread synchronization.
Added a ThreadManager to ease tracking thread creation/deletion, suspend and resume
of the main thread.

Added hooks for pthread_condition and signal.
This point is a bit tricky because there is two "pthread_cond_wait" version in
pthread lib, and you have to hook to the good one. Right now I manually selected
the version, which is bad (see libTAS/hook.h:46).