Reword first sentence of "Prevent writing to disk" option
"This option aims at preventing the game to save in savefiles on disk"
-> "This option aims to prevent the game from saving its savefiles on
disk".

Usually, people say "aims to" in this context, not "aims at". "aims at"
is used for like, pointing a physical gun at a target, or something, not
talking about semi-abstract goals. Furthermore, we say "from saving",
and not "to save" - one word is exclusively used for allowing an action,
the other is exclusively used for disallowing an action.

Then, "save in savefiles on disk" is technically correct, but it's far
easier to understand "saving its savefiles on disk".