Don't omit middle operand of ternary expression
Omitting the middle operand of a ternary expression is a GCC compiler
extension and is non-standard. This means the code can't otherwise be
compiled on Clang, for example.

The extension is equivalent to repeating the first operand (without its
implicit conversion to bool), except it doesn't double-evaluate.
However, in this case, the first expression is of a pointer, so it's
perfectly okay to double-evaluate it here.