Raise the minimum audio buffer size
The issue with mixing audio only one per frame is that with a too low
audio buffer, with have an underrun situation, e.g.:
buffer size at 1024 samples, games pushes groups of 512 samples, and
mixing takes 735 samples at 60 fps. Now if the buffer at currently 600
samples, the game cannot push a 512-samples group because it would be 
over the buffer size, but mixing cannot take place because it needs 735
samples.

For now I'm increasing the minimum buffer size. A cleaner way would be
to allow the game to push audio buffers *during* mixing, but it needs
a lot more work, because for now there's only a single mutex for all
audio, it would need a mutex per audio source, mix groups of
samples and unlocking the mutex in between.