Change the input editor toggling
Now input changes in the input editor are now performed on the main
thread instead of the UI thread (changes are pushed and processed by
main thread when available). The only except for now is input pasting
which is still performed by UI thread.

Visually, inputs that are still processing are showed with transparancy
effect (could be change to something else depending on feedback). The
main case when this is relevent is for rewinding, because state loading
takes a lot of time, and users can now still toggle inputs in the
meantime. It also support the case where state loading failed for some
reason, and input changes will correctly be discarded by main thread