Change parsing /proc/self/maps
Before, we were loading the entire file in memory, inside our
savestate-safe reserved memory. However, the file size is unbounded,
and it happens that Hollow Knight has a very fractionated memory
mapping, and the file can reach the 1 MB limit.

As such, we now make a copy of the file in a tmp file, and we parse
each line as needed.