Try to autodetect game libraries
Most of the time, games are shipped with needed libraries, and those
are located inside the game directory or a subdirectory. We detect if
there is a missing library from the output of `ld`, and lookup this
library inside the entire game directory. This is very simple but
should cover most of the cases.

Cases not covered-up: game executable in a subdirectory and libraries
inside another subdirectory (e.g. FTL), game with 32-bit and 64-bit
binaries, and both sets of libraries with identical names.