Make TLS internals weak
Starting from GLIBC 2.35, __libc_thread_freeres symbol is no longer
available to reset TLS, which is required for recycle threads feature.
So we make the internal symbols weak, and also warn the users that
recycle threads may not work. We don't prevent users from using the
feature, because it may still work fine.