Split timer into monotonic/realtime (#474)
There are different clock times that can be queried on Linux, which are
well described in clock_gettime() manpage. There are two main
categories:

* a monotonic clock that starts at an arbitrary value and only
  increases. It cannot be modified by the user
* a realtime clock whose value corresponds to the Unix timestamp, and
  which can be modified by the user

In one speedrun we modify the system clock to save time, so in
preparation for its support, we need to know which function returns
which category of clock time. We make our monotonic clock starts at 0,
so that it is easier to compute movie length and such. Realtime is the
one that is set by the user at startup. Both are shown on GUI main
window, and realtime also has the corresponding date and time.