Disable launching debugger if STDIN is not a TTY
This disables being able to use the "Launch with GDB" or "Launch with
LLDB" button if libTAS is not attached to a terminal.

Otherwise, a user who doesn't know any better could try to launch with a
debugger, and then get confused when they don't see anything noticeably
different - which is bad when closing the game doesn't close the
debugger, so if they did it over and over they'd have a ton of zombie
processes on their system.

At least one user in the libTAS Discord server has fallen victim to
this.