Change conditions for Unity threads (#515)
Before, we were storing the pointer of the thread start routine.
Teslagrad was soflocking during the intro cutscene because one of the
thread classified as loading thread by libTAS was not waiting and was
sleeping instead. This thread has the right start routine pointer, but
was started from a different address than the other threads.

So I changed the condition so that both the pointer of the start
routine thread and the pointer of the caller must match to classify
as a loading thread. This may cause desyncs in other Unity games, so
I might revert that change.