Immediately detach all created threads
We always had the problem of handling zombie threads. What I changed
here is that I detach all threads. This removes the issue of detaching
a thread twice after loading a savestate, and I can exclude the list
of zombie from the threads to monitor to accept loading a savestate.
This also fixes a softlock when trying to suspend a zombie thread,
because the underlying linux process may already have exited.