Use SA_SIGINFO / sa_sigaction for the checkpoint signal handler.
This ensures that the machine context of the signal will be saved and put somewhere where libTAS can save and restore it. This is typically not an actual issue, as long as the stack pointer is correct then on restore it will just return to after the saving signal raise. However, if the stack pointer is wrong (bad luck?), then it will just return to garbage and most likely crash immediately.