Add SDL_WINDOWEVENT_SIZE_CHANGED to the banned SDL2 events
This event seems to be able to just cause crashes (test case being Arma Cold War Assault, which triggers the event a few frames into execution and crashes due to it). I'm not sure if this is a proper fix, nor do I really understand why a crash occurs due to this event. It could be due to it invalidating the SDL GL Context (setting it to NULL) and that causing conflicts with libTAS's wrappers, maybe?