Implement AL_SOFT_buffer_sub_data support
This is require for Arma CWA for its dialog and music, and to avoid a crash due to AL_BYTE_RW_OFFSETS_SOFT not being responded to. The audio doesn't work for that game however, due to its heavy reliance on 3D audio functions which libTAS does not implement.
These changes are somewhat naive, as this extension is deprecated and no longer actually used as it conflicts with the AL_SOURCE_RADIUS_EXT extension (due to overlapping enum with AL_BYTE_RW_OFFSETS_SOFT). Although that bridge could be crossed when it comes to it.