Update hookpatch method
When patching a function to intercept, some instructions that are
transferred to the trampoline function may break, in particular those
that contain a relative address. Because instruction offsets are encoded
in 32-bit at most, we must ensure that the trampoline function lies
close to the patched method.

So we mmap a memory segment in the first 32-bit address, so that it will
be in reach with the executable memory (this was lazy and does not
support when a game would a base address outside the 32-bit range).
In this segment, we write the trampoline functions, and update each
instruction that contains a 32-bit relative address. For now, I only
implemented the instructions that were presents in some intercepted
real functions, but the code can be extended easily.