Use the correct ucontext definition for checkpoint signal handler
The ucontext here is allocated on the linux kernel size, and thus it does not necessarily have the same size as the libc ucontext
libc ucontext may be extended compared to the baseline linux ucontext, and thus the actual size of the passed ucontext might be smaller
In practice, this usually isn't an issue, as these are ultimately stack allocated (in the alt stack), and there's typically plenty of room available. Regardless, this still doesn't make it immune to possible issues.