Properly checkpoint floating point state
The floating point state in the ucontext is not in the ucontext directly, but rather it's in another buffer accessed by a pointer (libc extends ucontext to store it in the ucontext itself, but the kernel doesn't do this). In practice, this pointer probably doesn't change, since it's just a pointer to the alt stack, although saving/restoring the fp state regardless needs to be done via this pointer and it should remain unchanged when restoring the state.