Sync newer Unity engine games
For games that are shipped with `UnityPlayer_s.debug` file, we can
gather the address of `UnityClassic::Baselib_SystemFutex_Wait()`
function that is used by loading threads to wait for jobs, instead of
`sem_wait()` used by older engine versions.

We also use the given thread name to determine if a thread is a loading
thread.