Fix XCB event copying and injection
XCB event objects are expected to be allocated by malloc (not new!), and the caller is supposed to free them (which might not work for new!).
xcb_ge_generic_event_t events have variable length, so simply copying the first 36 bytes is not enough. This primarily shows issues with Xlib and Xcb being used together.