Remove some commented out code
This was kind of a TODO, to try to implement the reply side of these hooks, however it turns out to be really tricky, the only sensible thing seems to be to stache the sequence numbers returned then override replies using those sequence numbers. very annoying to do anyways, and it's not needed for now for Ruffle, so not bothering.