Find holes in shared mapped files for savestates
Shared mapped files or shared anonymous mappings (which have an
underlying file) can contain holes, for example by calling ftruncate()
on the file before mapping it, which reserves the memory but does not
allocate it. We can detect holes by opening the file and using lseek()
with `SEEK_DATA` to query the next data location in the file.

We can skip saving each memory page whose file contains a hole at the
specific location. When the mapped file is not available (anonymous,
or file was unlinked), `/proc/self/map_files` will still contain the
file.

However, that directory is only accessible if the process has
CAP_CHECKPOINT_RESTORE capability, which can be inherited by setting the
ambient capability in libTAS program. Users will need to run

`sudo setcap cap_checkpoint_restore+eip "$(which libTAS)"`

to enable it.